fixes https://github.com/jaymoulin/google-music-manager/issues/15 - deduplicate API

+++ google_music_manager_uploader/uploader_daemon.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
+# coding: utf-8
 
 import sys
 import time
@@ -8,6 +8,7 @@ import os
 import glob
 import netifaces
 import argparse
+import requests
 
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
@@ -18,24 +19,58 @@ __DEFAULT_IFACE__ = netifaces.gateways()['default'][netifaces.AF_INET][1]
 __DEFAULT_MAC__ = netifaces.ifaddresses(__DEFAULT_IFACE__)[netifaces.AF_LINK][0]['addr'].upper()
 
 
+class DeduplicateApi:
+    def __init__(self, uri: str) -> None:
+        self.uri = uri
+
+    def exists(self, file_path: str) -> bool:
+        result = requests.request('GET', self.uri + "/", data={"path": file_path})
+        return result.status_code == 200 or result.status_code == 204
+
+    def save(self, file_path: str) -> None:
+        requests.request('POST', self.uri + "/", data={"path": file_path})
+
+    def remove(self, file_path: str) -> None:
+        requests.request('DELETE', self.uri + "/", data={"path": file_path})
+
+
 class MusicToUpload(FileSystemEventHandler):
     def on_created(self, event) -> None:
         self.logger.info("Detected new files!")
         if os.path.isdir(self.path):
             files = [file for file in glob.glob(glob.escape(self.path) + '/**/*', recursive=True)]
             for file_path in files:
-                upload_file(self.api, file_path, self.logger, remove=self.remove)
+                upload_file(
+                    api=self.api,
+                    file_path=file_path,
+                    logger=self.logger,
+                    remove=self.remove,
+                    deduplicate_api=self.deduplicate_api,
+                )
         else:
-            upload_file(self.api, event.src_path, self.logger, remove=self.remove)
+            upload_file(
+                api=self.api,
+                file_path=event.src_path,
+                logger=self.logger,
+                remove=self.remove,
+                deduplicate_api=self.deduplicate_api
+            )
 
 
-def upload_file(api: Musicmanager, file_path: str, logger: logging.Logger, remove: bool = False) -> None:
+def upload_file(
+    api: Musicmanager,
+    file_path: str,
+    logger: logging.Logger,
+    remove: bool = False,
+    deduplicate_api: DeduplicateApi = None,
+) -> None:
     """
     Uploads a specific file by its path
-    :param api: Musicmanager object to upload file though
+    :param api: Musicmanager. object to upload file though
     :param file_path: Path to MP3 file to upload
     :param logger: logging.Logger object for logs
     :param remove: Boolean. should remove file? False by default
+    :param deduplicate_api: DeduplicateApi. Api for deduplicating uploads. Nano by default
     :raises CallFailure:
     :return:
     """
@@ -43,8 +78,15 @@ def upload_file(api: Musicmanager, file_path: str, logger: logging.Logger, remov
     while retry > 0:
         try:
             if os.path.isfile(file_path):
-                logger.info("Uploading : " + file_path)
+                logger.info("Uploading {}? " % file_path)
+                if deduplicate_api:
+                    exists = deduplicate_api.exists(file_path)
+                    logger.info("Deduplicate API: file exists? " + "yes" if exists else "no")
+                    if exists:
+                        return
                 uploaded, matched, not_uploaded = api.upload(file_path, True)
+                if (uploaded or matched) and deduplicate_api:
+                    deduplicate_api.save(file_path)
                 if remove and (uploaded or matched):
                     os.remove(file_path)
             retry = 0
@@ -64,7 +106,8 @@ def upload(
     oauth: str = os.environ['HOME'] + '/oauth',
     remove: bool = False,
     uploader_id: str = __DEFAULT_MAC__,
-    oneshot: bool = False
+    oneshot: bool = False,
+    deduplicate_api: str = None,
 ) -> None:
     logging.basicConfig(level=logging.INFO)
     logger = logging.getLogger(__name__)
@@ -75,6 +118,7 @@ def upload(
         print("Error with oauth credentials")
         sys.exit(1)
     observer = None
+    deduplicate = DeduplicateApi(deduplicate_api) if deduplicate_api else None
     if not oneshot:
         event_handler = MusicToUpload()
         event_handler.api = api
@@ -83,12 +127,13 @@ def upload(
         event_handler.path = directory
         event_handler.remove = remove
         event_handler.logger = logger
+        event_handler.deduplicate_api = deduplicate
         observer = Observer()
         observer.schedule(event_handler, directory, recursive=True)
         observer.start()
     files = [file for file in glob.glob(glob.escape(directory) + '/**/*', recursive=True)]
     for file_path in files:
-        upload_file(api, file_path, logger, remove=remove)
+        upload_file(api, file_path, logger, remove=remove, deduplicate_api=deduplicate)
     if oneshot:
         sys.exit(0)
     try:
@@ -101,28 +146,50 @@ def upload(
 
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument("--directory", '-d', default='.', help="Music Folder to upload from (default: .)")
+    parser.add_argument(
+        "--directory",
+        '-d',
+        default='.',
+        help="Music Folder to upload from (default: .)"
+    )
     parser.add_argument(
         "--oauth",
         '-a',
         default=os.environ['HOME'] + '/oauth',
         help="Path to oauth file (default: ~/oauth)"
     )
-    parser.add_argument("-r", "--remove", action='store_true', help="Remove files if present (default: False)")
+    parser.add_argument(
+        "-r",
+        "--remove",
+        action='store_true',
+        help="Remove files if present (default: False)"
+    )
     parser.add_argument(
         "--uploader_id",
         '-u',
         default=__DEFAULT_MAC__,
         help="Uploader identification (should be an uppercase MAC address) (default: <current eth0 MAC address>)"
     )
-    parser.add_argument("--oneshot", '-o', action='store_true', help="Upload folder and exit (default: False)")
+    parser.add_argument(
+        "--oneshot",
+        '-o',
+        action='store_true',
+        help="Upload folder and exit (default: False)"
+    )
+    parser.add_argument(
+        "--deduplicate_api",
+        '-w',
+        default=None,
+        help="Deduplicate API (should be HTTP and compatible with the manifest (see README)) (default: None)"
+    )
     args = parser.parse_args()
     upload(
         directory=args.directory,
         oauth=args.oauth,
         remove=args.remove,
         uploader_id=args.uploader_id,
-        oneshot=args.oneshot
+        oneshot=args.oneshot,
+        deduplicate_api=args.deduplicate_api
     )
